Performance: far LOD for OSM building shells (reflection and far cascades too), coarse far street ground, three-level crowd, shadow gates for small casters, tiled street props; dynamic resolution follows frame pacing on Metal, where ANGLE's GPU timer over-reads 2.5-6x

+++ src/world/osm/buildings/protocol.ts
@@ -17,10 +17,10 @@ export interface BuildingsRequest {
 
 export interface BuildingsResult {
   facade: MeshArrays;
-  /** Tile table of `facade` / `roof` (shared/mesh-tiles.ts tileIndex). */
-  facadeTiles: Float32Array;
+  /** Near / far leaf tables of `facade` / `roof` (shared/mesh-tiles.ts lodTileIndex). */
+  facadeTiles: Float64Array;
   roof: MeshArrays;
-  roofTiles: Float32Array;
+  roofTiles: Float64Array;
   details: Partial<Record<DetailKind, DetailStream>>;
   tiles: DetailTiles;
   /** INSTANCE_STRIDE records per rooftop prop kind. */
